Added a bunch more ALU instructions, e.g. bitwise operators.
Added basic support for streams, although currently only support streams of int type.
Introduced support for tasks, but there is a lot of work to be done there. In particular, need to move to thread pool model, but next up is the audio thread.
Added support for SDL and .Screen/XXX insructions.
Added support for G1, which is the mouse generator.